+++ src/main/java/net/silentchaos512/mechanisms/item/NoPlaceBucketItem.java
@@ -1,0 +1,26 @@
+package net.silentchaos512.mechanisms.item;
+
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.fluid.Fluid;
+import net.minecraft.item.BucketItem;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.ActionResultType;
+import net.minecraft.util.Hand;
+import net.minecraft.world.World;
+
+import java.util.function.Supplier;
+
+/**
+ * A bucket with right-click functionality removed. Used for fluids without blocks.
+ */
+public class NoPlaceBucketItem extends BucketItem {
+    public NoPlaceBucketItem(Supplier<? extends Fluid> supplier, Properties builder) {
+        super(supplier, builder);
+    }
+
+    @Override
+    public ActionResult<ItemStack> onItemRightClick(World worldIn, PlayerEntity playerIn, Hand handIn) {
+        return new ActionResult<>(ActionResultType.PASS, playerIn.getHeldItem(handIn));
+    }
+}